Create EDD Predictor
Pregnancy Expected Date of Delivery (EDD) and Postnatal care (PNC) at 6 weeks Predictor App is a mobile application designed to provide expectant mothers with an estimated date of delivery and when they are supposed to return with their newborns for postnatal care at 6 weeks based on various inputs and calculations. The app aims to assist pregnant women in tracking their pregnancy progress and better prepare for the arrival of their baby. It will leverage established medical algorithms and data to generate accurate predictions